chore: signup page refactoring - useSwipe hooks - fix MUI props
